# rest/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Chainspace REST API endpoints",
        "title": "Chainspace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/api/sbac/object": {
            "post": {
                "description": "TODO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sbac"
                ],
                "summary": "Create an object to a shard",
                "operationId": "createObject",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ObjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ObjectIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/sbac/states": {
            "get": {
                "description": "TODO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sbac"
                ],
                "summary": "Get the states",
                "operationId": "states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/sbac/transaction": {
            "post": {
                "description": "TODO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sbac"
                ],
                "summary": "Add a transaction to a shard",
                "operationId": "addTransaction",
                "parameters": [
                    {
                        "description": "transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/api/sbac/transaction-checked": {
            "post": {
                "description": "TODO",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sbac"
                ],
                "summary": "Add a transaction which is already checked to a shard",
                "operationId": "addTransaction",
                "parameters": [
                    {
                        "description": "transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.Dependency": {},
        "api.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.Object": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "value": {
                    "type": "object"
                },
                "versionId": {
                    "type": "string"
                }
            }
        },
        "api.ObjectIDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "api.ObjectRequest": {
            "type": "object",
            "properties": {
                "object": {
                    "type": "object"
                }
            }
        },
...
        "api.OutputObject": {
            "type": "object",
            "properties": {
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
        "api.Trace": {
            "type": "object",
            "properties": {
                "contractId": {
                    "type": "string"
                },
                "dependencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Dependency"
                    }
                },
                "inputObjectVersionIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "inputReferenceVersionIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "outputObjects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.OutputObject"
                    }
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "procedure": {
                    "type": "string"
                },
                "returns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.Transaction": {
            "type": "object",
            "properties": {
                "mappings": {
                    "type": "object"
                },
                "signatures": {
                    "type": "object"
                },
                "traces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Trace"
                    }
                }
            }
        },
...
    }
}